fix(githubrelease): account for "does not exist" error when there is a preset mismatch
BREAKING CHANGE:

If the preset is incorrect when calling `conventional-recommended-bump` `stderr`
will display a message that the *"<preset> does not exits"*. This may be the case
when a preset is not passed in as an argument and `conventional-commits-detector`
returns a different commit message style than what is being used. If this is the
case we reset the `preset` to `angular` and continue on. If the wrong preset is
present during the changelog edit process you will notice that there will be no
entries written for the current release. If this is the case you should enter a
present of your choosing:

```
githubrelease -P ember
```